cleanup layout files Migrated MediaInfoFragment2.kt to MediaInfoFragment.kt
