+++ app/src/main/java/com/afterroot/watchdone/ui/MainActivity.kt
@@ -347,12 +347,6 @@ class MainActivity : AppCompatActivity() {
                     binding.fab.hide()
                     drawerToggle.progress(0f, 1f) // As back arrow
                 }
-                R.id.navigation_media_info_2 -> {
-                    setTitle(null)
-                    binding.titleLayout.visible(false)
-                    binding.fab.hide()
-                    drawerToggle.progress(0f, 1f) // As back arrow
-                }
             }
         }
     }
