+++ ui/media/src/main/java/com/afterroot/watchdone/ui/media/adapter/MultiPagingAdapter.kt
@@ -14,15 +14,26 @@
  */
 package com.afterroot.watchdone.ui.media.adapter
 
+import android.content.Context
+import android.content.ContextWrapper
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.appcompat.widget.AppCompatImageView
+import androidx.core.view.updateLayoutParams
 import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.afterroot.utils.extensions.visible
+import com.afterroot.watchdone.base.GlideApp
+import com.afterroot.watchdone.binding.transitionOptions
+import com.afterroot.watchdone.data.model.Movie
+import com.afterroot.watchdone.data.model.TV
 import com.afterroot.watchdone.diff.MultiDiffCallback
 import com.afterroot.watchdone.media.databinding.ListItemMovieBinding
 import com.afterroot.watchdone.media.databinding.ListItemTvBinding
+import com.afterroot.watchdone.resources.R
 import com.afterroot.watchdone.settings.Settings
 import com.afterroot.watchdone.ui.common.ItemSelectedCallback
+import com.afterroot.watchdone.utils.getScreenWidth
 import info.movito.themoviedbapi.model.Multi
 
 class MultiPagingAdapter(private val callback: ItemSelectedCallback<Multi>, var settings: Settings) :
@@ -97,5 +108,75 @@ class MultiPagingAdapter(private val callback: ItemSelectedCallback<Multi>, var 
             },
             callback
         )
+    }
+}
+
+class MoviesListViewHolder(
+    val binding: ListItemMovieBinding,
+    val callback: ItemSelectedCallback<Multi>
+) : RecyclerView.ViewHolder(binding.root) {
+    private val posterView: AppCompatImageView = binding.poster
+    private val context: Context = (binding.root.context as ContextWrapper).baseContext
+    private var heightRatio: Float = 3f / 2f
+    private val width = context.getScreenWidth() / context.resources.getInteger(R.integer.grid_item_span_count)
+    fun bind(movie: Multi) {
+        binding.apply {
+            movieDb = movie as Movie
+            root.setOnClickListener {
+                callback.onClick(bindingAdapterPosition, root)
+                callback.onClick(bindingAdapterPosition, root, movie)
+            }
+            root.setOnLongClickListener {
+                callback.onLongClick(bindingAdapterPosition, movie)
+                return@setOnLongClickListener true
+            }
+            isWatched.visible(movie.isWatched)
+        }
+        posterView.updateLayoutParams {
+            this.width = this@MoviesListViewHolder.width
+            this.height = (width * heightRatio).toInt()
+        }
+
+        GlideApp.with(context).load(binding.baseUrl + binding.posterSize + binding.movieDb?.posterPath)
+            .override(width, (width * heightRatio).toInt())
+            .error(R.drawable.ic_placeholder_movie)
+            .transition(transitionOptions)
+            .centerCrop()
+            .into(posterView)
+    }
+}
+
+class TVListViewHolder(
+    val binding: ListItemTvBinding,
+    val callback: ItemSelectedCallback<Multi>
+) : RecyclerView.ViewHolder(binding.root) {
+    private val posterView: AppCompatImageView = binding.poster
+    private var heightRatio: Float = 3f / 2f
+    private val context: Context = (binding.root.context as ContextWrapper).baseContext
+    private val width = context.getScreenWidth() / context.resources.getInteger(R.integer.grid_item_span_count)
+    fun bind(tv: Multi) {
+        binding.apply {
+            tvSeries = tv as TV
+            root.setOnClickListener {
+                callback.onClick(bindingAdapterPosition, root)
+                callback.onClick(bindingAdapterPosition, root, tv)
+            }
+            root.setOnLongClickListener {
+                callback.onLongClick(bindingAdapterPosition, tv)
+                return@setOnLongClickListener true
+            }
+            isWatched.visible(tv.isWatched)
+        }
+        posterView.updateLayoutParams {
+            this.width = this@TVListViewHolder.width
+            this.height = (width * heightRatio).toInt()
+        }
+
+        GlideApp.with(context).load(binding.baseUrl + binding.posterSize + binding.tvSeries?.posterPath)
+            .override(width, (width * heightRatio).toInt())
+            .error(R.drawable.ic_placeholder_tv)
+            .transition(transitionOptions)
+            .centerCrop()
+            .into(posterView)
     }
 }
