--- ui/media/src/main/java/com/afterroot/watchdone/ui/media/adapter/delegate/AdapterType.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (C) 2020-2022 Sandip Vaghela
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.afterroot.watchdone.ui.media.adapter.delegate
-
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import info.movito.themoviedbapi.model.Multi
-
-interface AdapterType {
-    fun onCreateViewHolder(parent: ViewGroup): RecyclerView.ViewHolder
-    fun onBindViewHolder(holder: RecyclerView.ViewHolder, item: Multi)
-}
