--- ui/media/src/main/res/layout/list_item_cast.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="personDetail"
-            type="info.movito.themoviedbapi.model.people.PersonCast" />
-
-        <variable
-            name="personCrew"
-            type="info.movito.themoviedbapi.model.people.PersonCrew" />
-    </data>
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="?attr/selectableItemBackground"
-        android:clipToPadding="true"
-        app:cardCornerRadius="@dimen/card_radius"
-        app:cardUseCompatPadding="true">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="?attr/selectableItemBackground">
-
-            <androidx.appcompat.widget.AppCompatImageView
-                android:id="@+id/cast_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0.0"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                tools:src="@tools:sample/avatars" />
-
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/cast_name"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:paddingStart="@dimen/padding_horizontal_half"
-                android:paddingLeft="@dimen/padding_horizontal_half"
-                android:paddingTop="4dp"
-                android:paddingEnd="@dimen/padding_horizontal_half"
-                android:paddingRight="@dimen/padding_horizontal_half"
-                android:text="@{personDetail.name ?? personCrew.name}"
-                android:textAppearance="@style/TextAppearance.WatchDone.BodySmall"
-                app:layout_constraintBottom_toTopOf="@+id/character_name"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/cast_iv"
-                app:layout_constraintVertical_chainStyle="packed"
-                tools:text="@tools:sample/full_names" />
-
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/character_name"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="1dp"
-                android:paddingStart="@dimen/padding_horizontal_half"
-                android:paddingLeft="@dimen/padding_horizontal_half"
-                android:paddingEnd="@dimen/padding_horizontal_half"
-                android:paddingRight="@dimen/padding_horizontal_half"
-                android:paddingBottom="@dimen/padding_vertical_half"
-                android:text="@{personDetail.character ?? personCrew.job}"
-                android:textAppearance="@style/TextAppearance.WatchDone.BodySmall"
-                android:textStyle="italic"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/cast_name"
-                app:layout_constraintVertical_chainStyle="packed"
-                tools:text="@tools:sample/full_names" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.cardview.widget.CardView>
-</layout>
