--- ui/media/src/main/res/layout/list_item_person.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="personDetail"
-            type="info.movito.themoviedbapi.model.people.Person" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="?attr/selectableItemBackground">
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/cast_iv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/padding_normal_half"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:src="@tools:sample/avatars" />
-
-        <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/cast_name"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:maxLines="2"
-            android:paddingStart="@dimen/padding_horizontal_half"
-            android:paddingLeft="@dimen/padding_horizontal_half"
-            android:paddingTop="4dp"
-            android:paddingEnd="@dimen/padding_horizontal_half"
-            android:paddingRight="@dimen/padding_horizontal_half"
-            android:text="@{personDetail.name}"
-            android:textAppearance="@style/TextAppearance.WatchDone.BodySmall"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cast_iv"
-            app:layout_constraintVertical_chainStyle="packed"
-            tools:text="@tools:sample/full_names" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
